+++ run/jobs/queueMonitoring.js
@@ -2,23 +2,39 @@ const { Queue } = require('bullmq');
 const connection = require('../lib/redis');
 const logger = require('../lib/logger');
 const { createIncident } = require('../lib/opsgenie');
-const { queueMonitoringMaxProcessingTime, queueMonitoringHighProcessingTimeThreshold, queueMonitoringHighWaitingJobCountThreshold, queueMonitoringMaxWaitingJobCount } = require('../lib/env');
+const { maxTimeWithoutEnqueuedJob, queueMonitoringMaxProcessingTime, queueMonitoringHighProcessingTimeThreshold, queueMonitoringHighWaitingJobCountThreshold, queueMonitoringMaxWaitingJobCount } = require('../lib/env');
 
-const monitoredQueues = [
+const monitoredPerformances = [
     'blockSync', 'receiptSync'
 ];
 
+const monitoredActivity = ['blockSync'];
+
 module.exports = async () => {
-    if (!queueMonitoringMaxProcessingTime() || !queueMonitoringHighProcessingTimeThreshold() || !queueMonitoringHighWaitingJobCountThreshold() || !queueMonitoringMaxWaitingJobCount()) {
-        return logger.info('Queue monitoring is not enabled');
+    if (!maxTimeWithoutEnqueuedJob() || !queueMonitoringMaxProcessingTime() || !queueMonitoringHighProcessingTimeThreshold() || !queueMonitoringHighWaitingJobCountThreshold() || !queueMonitoringMaxWaitingJobCount()) {
+        logger.info('Queue monitoring is not enabled');
+        return;
     }
 
-    for (const queueName of monitoredQueues) {
+    let incidentCreated = false;
+
+    for (const queueName of monitoredActivity) {
+        const queue = new Queue(queueName, { connection });
+        const completedJobs = await queue.getCompleted();
+        const latestJob = completedJobs[0];
+
+        if (latestJob.timestamp < Date.now() - maxTimeWithoutEnqueuedJob() * 1000) {
+            await createIncident(`${queueName} queue issue (no jobs enqueued)`, `Latest job timestamp: ${new Date(latestJob.timestamp).toISOString()}`);
+            incidentCreated = true;
+        }
+    }
+
+    for (const queueName of monitoredPerformances) {
         const queue = new Queue(queueName, { connection });
         const completedJobs = await queue.getCompleted();
 
         const averageProcessingTime = completedJobs.reduce((a, b) => {
-            return b.finishedOn ? a + (b.finishedOn - b.processedOn) / 1000 : a;
+            return b && b.finishedOn ? a + (b.finishedOn - b.processedOn) / 1000 : a;
         }, 0) / completedJobs.length;
 
         const waitingJobCount = await queue.getWaitingCount();
@@ -28,12 +44,13 @@ module.exports = async () => {
 
         if (
             averageProcessingTime > queueMonitoringMaxProcessingTime() ||
-            (averageProcessingTime > queueMonitoringHighProcessingTimeThreshold() && waitingJobCount >= queueMonitoringHighWaitingJobCountThreshold()) ||
-            waitingJobCount >= queueMonitoringMaxWaitingJobCount()
+            waitingJobCount >= queueMonitoringMaxWaitingJobCount() ||
+            (averageProcessingTime >= queueMonitoringHighProcessingTimeThreshold() && waitingJobCount >= queueMonitoringHighWaitingJobCountThreshold())
         ) {
-            await createIncident(`${queueName} Queue monitoring`,
-                `Waiting job count: ${waitingJobCount} - Delayed job count: ${delayedJobCount} - Average processing time: ${averageProcessingTime}s`
-            );
+            await createIncident(`${queueName} queue issue (performance)`, `Waiting job count: ${waitingJobCount} - Delayed job count: ${delayedJobCount} - Average processing time: ${averageProcessingTime}s`);
+            incidentCreated = true;
         }
     }
+
+    return incidentCreated;
 };