+++ run/lib/rpc.js
@@ -73,7 +73,12 @@ const getBalanceChange = async (address, token, blockNumber, rpcServer) => {
     };
 }
 
+let providers = {};
+
 const getProvider = function(url) {
+    if (providers[url])
+        return providers[url];
+
     const rpcServer = new URL(url);
 
     let provider;
@@ -92,7 +97,8 @@ const getProvider = function(url) {
             password: rpcServer.password
         };
 
-    return new provider(authenticatedUrl);
+    providers[url] = new provider(authenticatedUrl);
+    return providers[url];
 };
 
 class DexFactoryConnector {
@@ -208,7 +214,7 @@ class ProviderConnector {
 
 class Tracer {
 
-    #ERRORS_TO_IGNORE = [-32601];
+    #ERRORS_TO_IGNORE = [-32601, -32000];
     #bytecodes = {};
 
     constructor(server, db, type = 'other') {
@@ -229,7 +235,8 @@ class Tracer {
 
         if (error.error && this.#ERRORS_TO_IGNORE.includes(error.error.code))
             return this.error = {
-                message: `Error code "${error.error.code}".`,
+                code: `Error code "${error.error.code}".`,
+                message: error.error.message,
                 error: error
             };