+++ run/scheduler.js
@@ -5,11 +5,12 @@ const RPC_HEALTH_CHECK_INTERVAL = 5 * 60 * 1000;
 const SUBSCRIPTION_CHECK_INTERVAL = 5 * 60 * 1000;
 const QUEUE_MONITORING_INTERVAL = 60 * 1000;
 const CANCEL_DEMO_INTERVAL = 60 * 60 * 1000;
+const BLOCK_SYNC_MONITORING_INTERVAL = 60 * 1000;
 
 (async () => {
     await enqueue(
-        'cancelDemoExplorers',
-        'cancelDemoExplorers',
+        'removeExpiredExplorers',
+        'removeExpiredExplorers',
         {},
         10,
         { every: CANCEL_DEMO_INTERVAL }
@@ -53,5 +54,13 @@ const CANCEL_DEMO_INTERVAL = 60 * 60 * 1000;
         {},
         10,
         { every: QUEUE_MONITORING_INTERVAL }
+    );
+
+    await enqueue(
+        'blockSyncMonitoring',
+        'blockSyncMonitoring',
+        {},
+        10,
+        { every: BLOCK_SYNC_MONITORING_INTERVAL }
     );
 })();