+++ run/tests/mocks/lib/opsgenie.js
@@ -1,0 +1,3 @@
+jest.mock('../../../lib/opsgenie', () => ({
+    createIncident: jest.fn().mockResolvedValue('OK')
+}));